requirements.txt
# Core Dependencies
Flask==3.0.0
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0

# Data Processing
pandas==2.1.4
numpy==1.26.3

# NLP and Sentiment Analysis
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1
vaderSentiment==3.3.2
transformers==4.36.2
torch==2.1.2

# Database
psycopg2-binary==2.9.9
pymongo==4.6.1
sqlalchemy==2.0.25
alembic==1.13.1

# Caching and Message Queue
redis==5.0.1
celery==5.3.4

# API and Web Scraping
requests==2.31.0
tweepy==4.14.0
feedparser==6.0.10
beautifulsoup4==4.12.2

# Dashboard and Visualization
dash==2.14.2
plotly==5.18.0
dash-bootstrap-components==1.5.0

# Utilities
pyyaml==6.0.1
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.23.3
httpx==0.27.2

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
